Fixes issue with mutual information function
Although MI was being calculated iteratively to be efficient, was not considering the case of multiple same valued candidates, which though in different positions in the sorted array - produce the same split obviously. Changed the code to fix this, by iterating the sorted array in the opposite order, and skipping calculating gain for identical elements.